third_party/blink/web_tests/fast/forms/calendar-picker/date-picker-with-enter-key.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>

<input type=date id=date value="2019-09-10">

<script>

test(() => {
  document.getElementById('date').focus();
  eventSender.keyDown("Enter");
  assert_equals(internals.pagePopupWindow, null);
}, "The date picker should not open with the Enter key.");

</script>
</body>
</html>
